# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Yuque1 API
  description: |
    复刻语雀核心子集 HTTP API 契约（SSOT）。

    **对齐文档**
    - Design-00 v1.2 总体技术方案
    - Design-01 v1.1 领域模型与数据库

    **约定**
    - 前缀：`/api/v1`
    - 鉴权：HttpOnly Cookie `sid`（同站）；分享读接口除外
    - 统一响应包络：`success` / `data` / `error` / `requestId`
    - 业务错误使用语义化 HTTP 状态码（禁止一律 200）
    - 资源 ID：ULID 字符串（26 字符）
    - 正文：`nodeId`；overwrite 必须带 `baseVersion`
    - 正文大小：UTF-8 **字节** ≤ 2097152（超限 413）；勿用 maxLength 模拟字节
    - 浏览器客户端：需登录请求使用 credentials/include（withCredentials）
    - 删除类接口统一 200 + 包络（不用 204）
    - 列表集合字段统一 `data.items`（含树扁平列表）
  version: 0.1.9
  contact:
    name: yuque1
servers:
  - url: /api/v1
    description: 相对当前站点（Nginx 同站反代）
...
paths:
...
  /auth/sms/send:
    post:
      tags: [Auth]
      summary: 发送短信验证码
      operationId: sendSmsCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSmsRequest'
      responses:
        '200':
          description: 已接受发送（不泄露用户是否已注册的细节策略由实现决定，默认统一成功文案）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeEmpty'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/sms/login:
    post:
      tags: [Auth]
      summary: 手机号验证码登录/注册
      description: |
        验证成功后创建 Session，并通过 `Set-Cookie: sid=...; HttpOnly; Path=/; SameSite=Lax; Secure(prod)` 下发。
        新用户同事务创建 User + Workspace。
        错误码：400 `SMS_CODE_INVALID` / `VALIDATION_ERROR`；429 `SMS_RATE_LIMITED` / `AUTH_LOCKED`。
      operationId: smsLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsLoginRequest'
      responses:
        '200':
          description: 登录成功
          headers:
            Set-Cookie:
              description: 会话 Cookie `sid`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      operationId: logout
      responses:
        '200':
          description: 已清除服务端 Session 与 Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeEmpty'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: 当前用户
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Auth]
      summary: 更新当前用户资料
      description: |
        支持 `nickname`（trim 后 1–64）与可选 `email`（无验证码；null/空清除）。
        至少提供其中一个字段。写入 audit `user.update_profile`。
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /nodes/{nodeId}/content/revisions:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    get:
      tags: [Content]
      summary: 列出覆盖前正文快照
      description: |
        成员可读。仅 overwrite 路径写入的 `content_revisions`（reason=overwrite_on_conflict）。
        列表不含 bodyMd，最多 50 条，按 createdAt 倒序。
      operationId: listNodeContentRevisions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeRevisionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /nodes/{nodeId}/content/revisions/{revisionId}:
    parameters:
      - $ref: '#/components/parameters/NodeId'
      - name: revisionId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Ulid'
    get:
      tags: [Content]
      summary: 获取单条覆盖快照（含正文）
      description: 成员可读；revision 必须属于该 node。
      operationId: getNodeContentRevision
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeRevision'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /nodes/{nodeId}/content:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    get:
      tags: [Content]
      summary: 获取文档正文
      description: |
        仅 type=doc。folder 或其它不可访问情况统一 404 NOT_FOUND（模糊策略）。
      operationId: getNodeContent
      responses:
        '200':
          description: OK（仅 type=doc）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Content]
      summary: 保存正文（乐观锁）
      description: |
        必须携带 `expectedVersion`；不匹配返回 409 DOC_VERSION_CONFLICT。
        正文 UTF-8 **字节**超过 2MB → 413。
        node 非 doc 时 404 NOT_FOUND。
      operationId: putNodeContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutContentRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeContentMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /nodes/{nodeId}/content/overwrite:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    post:
      tags: [Content]
      summary: 冲突后强制覆盖
      description: |
        必须携带 `baseVersion`（冲突面板中的 serverVersion）。
        成功：写入 content_revisions + 条件更新 + 审计。
        baseVersion 已过期 → 409。
      operationId: overwriteNodeContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverwriteContentRequest'
      responses:
        '200':
          description: 覆盖成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeContentMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /nodes/{nodeId}/content/save-as:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    post:
      tags: [Content]
      summary: 冲突后另存为新文档
      description: 不修改源文档；新建 doc 节点 + 正文 version=1。
      operationId: saveAsNodeContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAsContentRequest'
      responses:
        '201':
          description: 已创建副本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelopeNodeWithContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Redis Session ID
  parameters:
...
    NodeId:
      name: nodeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Ulid'
  schemas:
    Ulid:
      type: string
      minLength: 26
      maxLength: 26
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      description: ULID（Crockford Base32）
      example: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
    ErrorCode:
      type: string
      description: 业务错误码（实现与前端以此枚举为准）
      enum:
        - UNAUTHORIZED
        - NOT_FOUND
        - VALIDATION_ERROR
        - SMS_RATE_LIMITED
        - SMS_CODE_INVALID
        - AUTH_LOCKED
        - RATE_LIMITED
        - DOC_VERSION_CONFLICT
        - NODE_HAS_CHILDREN
        - NODE_CYCLE_DETECTED
        - PAYLOAD_TOO_LARGE
        - SHARE_INVALID
        - SERVICE_UNAVAILABLE
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, data, error, requestId]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
          description: 恒为 null
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          type: string
    User:
      type: object
      required: [id, mobileE164, nickname, mobileVerified]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        mobileE164:
          type: string
          pattern: '^\+[1-9]\d{6,14}$'
          example: '+8613800138000'
        nickname:
          type: string
        mobileVerified:
          type: boolean
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
...
    NodeType:
      type: string
      enum: [folder, doc]
    TreeNode:
      type: object
      required: [id, knowledgeBaseId, type, title, sortOrder, parentId, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        knowledgeBaseId:
          $ref: '#/components/schemas/Ulid'
        parentId:
          allOf:
            - $ref: '#/components/schemas/Ulid'
          nullable: true
        type:
          $ref: '#/components/schemas/NodeType'
        title:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserBrief:
      type: object
      required: [id, nickname]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        nickname:
          type: string
    DocumentContent:
      type: object
      required: [nodeId, bodyMd, version, updatedAt]
      properties:
        nodeId:
          $ref: '#/components/schemas/Ulid'
        bodyMd:
          type: string
          description: Markdown 原文
        version:
          type: integer
          minimum: 1
          description: content_version
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          allOf:
            - $ref: '#/components/schemas/UserBrief'
          nullable: true
    ContentMeta:
      type: object
      required: [nodeId, version, updatedAt]
      properties:
        nodeId:
          $ref: '#/components/schemas/Ulid'
        version:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
    ContentRevisionBrief:
      type: object
      required: [id, version, reason, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        version:
          type: integer
          minimum: 1
          description: 覆盖前 content_version（= baseVersion）
        reason:
          type: string
          enum: [overwrite_on_conflict]
        createdBy:
          allOf:
            - $ref: '#/components/schemas/UserBrief'
          nullable: true
        createdAt:
          type: string
          format: date-time
    ContentRevision:
      allOf:
        - $ref: '#/components/schemas/ContentRevisionBrief'
        - type: object
          required: [nodeId, bodyMd]
          properties:
            nodeId:
              $ref: '#/components/schemas/Ulid'
            bodyMd:
              type: string
              description: 覆盖前 Markdown 正文
...
    SendSmsRequest:
      type: object
      required: [mobileE164]
      properties:
        mobileE164:
          type: string
          description: E.164
          pattern: '^\+[1-9]\d{6,14}$'
          example: '+8613800138000'
    SmsLoginRequest:
      type: object
      required: [mobileE164, code]
      properties:
        mobileE164:
          type: string
          pattern: '^\+[1-9]\d{6,14}$'
          description: E.164
        code:
          type: string
          pattern: '^\d{6}$'
          minLength: 6
          maxLength: 6
          description: 6 位数字验证码
        nickname:
          type: string
          maxLength: 64
          description: 仅新用户注册时可选
    UpdateMeRequest:
      type: object
      minProperties: 1
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 64
          description: 展示昵称；服务端 trim 后校验非空
        email:
          type: string
          maxLength: 255
          nullable: true
          description: 可选邮箱；null 或空字符串清除；不验证所有权
...
    PutContentRequest:
      type: object
      required: [expectedVersion, bodyMd]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        bodyMd:
          type: string
          description: |
            Markdown，UTF-8。服务端按 **字节** 限制 byte_length ≤ 2097152；
            超限返回 413 PAYLOAD_TOO_LARGE（OpenAPI maxLength 不表示字节合同）。
    OverwriteContentRequest:
      type: object
      required: [baseVersion, bodyMd]
      properties:
        baseVersion:
          type: integer
          minimum: 1
          description: 冲突时的 serverVersion
        bodyMd:
          type: string
          description: |
            Markdown，UTF-8。服务端按 **字节** 限制 byte_length ≤ 2097152；
            超限返回 413 PAYLOAD_TOO_LARGE（OpenAPI maxLength 不表示字节合同）。
    SaveAsContentRequest:
      type: object
      required: [bodyMd]
      properties:
        bodyMd:
          type: string
          description: |
            Markdown，UTF-8。服务端按 **字节** 限制 byte_length ≤ 2097152；
            超限返回 413 PAYLOAD_TOO_LARGE（OpenAPI maxLength 不表示字节合同）。
        title:
          type: string
          maxLength: 512
          description: 默认「{原标题} (冲突副本)」
        parentId:
          allOf:
            - $ref: '#/components/schemas/Ulid'
          nullable: true
          description: 默认与源节点同父
...
    VersionConflictDetails:
      type: object
      required: [serverVersion]
      properties:
        serverVersion:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          allOf:
            - $ref: '#/components/schemas/UserBrief'
          nullable: true
    ErrorEnvelopeVersionConflict:
      type: object
      required: [success, data, error, requestId]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
          description: 恒为 null
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum: [DOC_VERSION_CONFLICT]
            message:
              type: string
            details:
              $ref: '#/components/schemas/VersionConflictDetails'
        requestId:
          type: string

    # --- Envelopes ---
    SuccessEnvelopeEmpty:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data: { nullable: true }
        error: { nullable: true }
        requestId: { type: string }
...
    SuccessEnvelopeUser:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          $ref: '#/components/schemas/User'
        error: { nullable: true }
        requestId: { type: string }
...
    SuccessEnvelopeContent:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          $ref: '#/components/schemas/DocumentContent'
        error: { nullable: true }
        requestId: { type: string }
    SuccessEnvelopeContentMeta:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          $ref: '#/components/schemas/ContentMeta'
        error: { nullable: true }
        requestId: { type: string }
    SuccessEnvelopeRevisionList:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ContentRevisionBrief'
        error: { nullable: true }
        requestId: { type: string }
    SuccessEnvelopeRevision:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          $ref: '#/components/schemas/ContentRevision'
        error: { nullable: true }
        requestId: { type: string }
    SuccessEnvelopeNodeWithContent:
      type: object
      required: [success, data, error, requestId]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          required: [node, content]
          properties:
            node:
              $ref: '#/components/schemas/TreeNode'
            content:
              $ref: '#/components/schemas/DocumentContent'
        error: { nullable: true }
        requestId: { type: string }
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            error:
              code: VALIDATION_ERROR
              message: 请求参数不合法
              details: {}
            requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            error:
              code: UNAUTHORIZED
              message: 未登录
            requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
    NotFound:
      description: 不存在或无权限（模糊 404）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            error:
              code: NOT_FOUND
              message: 资源不存在
            requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
...
    VersionConflict:
      description: 正文版本冲突（error.details 见 VersionConflictDetails）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeVersionConflict'
          example:
            success: false
            data: null
            error:
              code: DOC_VERSION_CONFLICT
              message: 文档已被他人更新
              details:
                serverVersion: 4
                updatedAt: '2026-07-12T10:00:00.000Z'
                updatedBy:
                  id: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
                  nickname: Alice
            requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
    PayloadTooLarge:
      description: 正文超过 2MB
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            error:
              code: PAYLOAD_TOO_LARGE
              message: 文档内容超过 2MB 限制
            requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
    RateLimited:
      description: 限流
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            sms:
              value:
                success: false
                data: null
                error:
                  code: SMS_RATE_LIMITED
                  message: 发送过于频繁
                requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
            generic:
              value:
                success: false
                data: null
                error:
                  code: RATE_LIMITED
                  message: 请求过于频繁
                requestId: '01ARZ3NDEKTSV4RRFFQ69G5FAV'
...
security:
  - cookieAuth: []